refactor(memory): split verb flows out of command wiring
memory.ts was 207 lines (>200 bar); extract the six run* verb flows and
their helpers into memory-verbs.ts, leaving memory.ts as thin commander
wiring. No behavior change.
